add service

+++ app/Http/Controllers/LogController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 
 use App\EmailLog;
 use App\CustomerLog;
+use App\ServiceLog;
 use Carbon\Carbon;
 
 class LogController extends Controller
@@ -31,11 +32,11 @@ class LogController extends Controller
     	$filler_date_from = $request->get('filler_date_from');
         $filler_date_to = $request->get('filler_date_to');
 
-        $filler_date_from = (!$filler_date_from) ? date('Y-m-d 00:00:00') : Carbon::createFromFormat('d-m-Y', $filler_date_from)->format('Y-m-d');
-        $filler_date_to = (!$filler_date_to) ? date('Y-m-d 23:59:59') : Carbon::createFromFormat('d-m-Y', $filler_date_to)->format('Y-m-d');
-
-    	$logs = CustomerLog::whereDate('created_at', '>', $filler_date_from)
-                                ->whereDate('created_at', '<', $filler_date_to)->get();
+        $filler_date_from = (!$filler_date_from) ? date('Y-m-d 00:00:00') : Carbon::createFromFormat('d-m-Y H:i:s', $filler_date_from . ' 00:00:00')->format('Y-m-d');
+        $filler_date_to = (!$filler_date_to) ? date('Y-m-d 23:59:59') : Carbon::createFromFormat('d-m-Y H:i:s', $filler_date_to . ' 23:59:59')->format('Y-m-d');
+        
+    	$logs = CustomerLog::whereDate('created_at', '>=', $filler_date_from)
+                                ->whereDate('created_at', '<=', $filler_date_to)->get();
 
         return view('log.index', [
         	'logs' => $logs,
@@ -51,6 +52,43 @@ class LogController extends Controller
             $id = $input['id'];
             if($id) {
                 $log = CustomerLog::where(['id' => $id])->first();
+                $log->delete();
+                return response(['success' => true]);
+            }
+            return response(['success' => false, 'message' => 'Không tìm thấy log']);
+        } 
+    }
+
+    /**
+     * Show the application dashboard.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index2(Request $request)
+    {
+    	$filler_date_from = $request->get('filler_date_from');
+        $filler_date_to = $request->get('filler_date_to');
+
+        $filler_date_from = (!$filler_date_from) ? date('Y-m-d 00:00:00') : Carbon::createFromFormat('d-m-Y H:i:s', $filler_date_from . ' 00:00:00')->format('Y-m-d');
+        $filler_date_to = (!$filler_date_to) ? date('Y-m-d 23:59:59') : Carbon::createFromFormat('d-m-Y H:i:s', $filler_date_to . ' 23:59:59')->format('Y-m-d');
+        
+    	$logs = ServiceLog::whereDate('created_at', '>=', $filler_date_from)
+                                ->whereDate('created_at', '<=', $filler_date_to)->get();
+        
+        return view('log.index2', [
+        	'logs' => $logs,
+        	'filler_date_from' => $filler_date_from,
+        	'filler_date_to' => $filler_date_to,
+        ]);
+    }
+
+    public function delete2(Request $request) 
+    {
+        if($request->ajax()) {
+            $input = $request->all();
+            $id = $input['id'];
+            if($id) {
+                $log = ServiceLog::where(['id' => $id])->first();
                 $log->delete();
                 return response(['success' => true]);
             }
